requirements.txt
feedparser>=6.0.11
python-dateutil>=2.8.2
google-genai>=0.3.0

